requirements.txt
python-decouple
mysqlclient
django
djangorestframework
firebase-admin
